requirements.txt
# Django core
Django>=5.0,<5.1
djangorestframework>=3.14.0

# Authentication
django-allauth>=0.61.0
dj-rest-auth>=5.0.0

# OAuth support
PyJWT>=2.8.0
requests>=2.31.0
cryptography>=41.0.0

# Production server
gunicorn>=21.2.0

# Database (production)
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# Static files
whitenoise>=6.6.0

# Environment variables
python-dotenv>=1.0.0

# CORS
django-cors-headers>=4.3.0
